More limitations for blocked users

+++ grid-paint/pages/actions.py
@@ -205,6 +205,10 @@ class ActionDeleteImage(BasicRequestHandler):
         artwork_id=self.request.get('id')
         artwork=dao.get_artwork(artwork_id)
         if self.user_info.superadmin or artwork.author_email==self.user_info.user_email:
+            if self.user_info.read_only:
+                self.response.set_status(403)
+                return            
+            
             comments=db.Comment.all().filter('artwork_ref =', artwork)
             for comment in comments:
                 comment.delete()
